feat: auth with token
