+++ packages/salesforcedx-vscode-core/package.nls.ja.json
@@ -2,6 +2,7 @@
   "running_tasks_title_text": "実行中のタスク",
   "force_auth_web_login_authorize_dev_hub_text": "SFDX: Dev Hub を認証",
   "force_auth_web_login_authorize_org_text": "SFDX: 組織を認証",
+  "force_auth_access_token_authorize_org_text": "SFDX: Authorize an Org using Session ID",
   "force_auth_logout_all_text": "SFDX: すべての認証済み組織からログアウト",
   "force_org_create_default_scratch_org_text": "SFDX: デフォルトのスクラッチ組織を作成...",
   "force_org_open_default_scratch_org_text": "SFDX: デフォルトの組織を開く",
