+++ packages/salesforcedx-vscode-core/package.nls.json
@@ -2,6 +2,7 @@
   "running_tasks_title_text": "Running Tasks",
   "force_auth_web_login_authorize_dev_hub_text": "SFDX: Authorize a Dev Hub",
   "force_auth_web_login_authorize_org_text": "SFDX: Authorize an Org",
+  "force_auth_access_token_authorize_org_text": "SFDX: Authorize an Org using Session ID",
   "force_auth_logout_all_text": "SFDX: Log Out from All Authorized Orgs",
   "force_org_create_default_scratch_org_text": "SFDX: Create a Default Scratch Org...",
   "force_org_open_default_scratch_org_text": "SFDX: Open Default Org",
